Fix RISC OS warning (try 2)
git-svn-id: file:///home/survex-svn/survex/trunk@1692 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/cvrotgfx.c
@@ -699,7 +699,7 @@ cvrotgfx_init(void)
    /* limits for OS routines (used for text) */
    xcMacOS = xcMac << eigX;
    ycMacOS = ycMac << eigY;
-   y_stretch *= (double)(1ul << eigX) / (1ul << eigY);
+   y_stretch *= (double)(1 << eigX) / (double)(1 << eigY);
    /* set origin to centre of screen +/- adjustment for labels */
    {
       int v;
